# Merge `Dry::Schema::Step::Scoped` into `Dry::Schema::Step`