--- a/lib/dry/schema/step.rb
+++ b/lib/dry/schema/step.rb
@@ -7,6 +7,9 @@ module Dry
   module Schema
     # @api private
     class Step
+      EMPTY_PATH = Path.new([]).freeze
+      private_constant :EMPTY_PATH
+
       # @api private
       attr_reader :name
 
@@ -17,53 +20,34 @@ module Dry
       attr_reader :executor
 
       # @api private
-      class Scoped
-        # @api private
-        attr_reader :path
-
-        # @api private
-        attr_reader :step
-
-        # @api private
-        def initialize(path, step)
-          @path = Path[path]
-          @step = step
-        end
-
-        # @api private
-        def scoped(new_path)
-          self.class.new(Path[[*new_path, *path]], step)
-        end
-
-        # @api private
-        def call(result)
-          result.at(path) do |scoped_result|
-            output = step.(scoped_result).to_h
-            target = Array(path)[0..-2].reduce(result) { |a, e| a[e] }
-
-            target.update(path.last => output)
-          end
-        end
-      end
+      attr_reader :path
 
       # @api private
-      def initialize(type:, name:, executor:)
+      def initialize(type:, name:, executor:, path: EMPTY_PATH)
         @type = type
         @name = name
         @executor = executor
+        @path = path
         validate_name(name)
       end
 
       # @api private
       def call(result)
-        output = executor.(result)
-        result.replace(output) if output.is_a?(Hash)
+        scoped_result = path.equal?(EMPTY_PATH) ? result : result.at(path)
+
+        output = executor.(scoped_result)
+        scoped_result.replace(output) if output.is_a?(Hash)
         output
       end
 
       # @api private
-      def scoped(path)
-        Scoped.new(path, self)
+      def scoped(parent_path)
+        self.class.new(
+          type: type,
+          name: name,
+          executor: executor,
+          path: Path.new([*parent_path, *path])
+        )
       end
 
       private