Second extract from https://github.com/dry-rb/dry-schema/pull/353

It makes things more streamlined. And preparational commit:
```
   Drop `Dry::Schema::Result#results`
    
    It's unused, we're working only with result_ast and if value coercer
    is not run on the Result (will be the case in the next commit) — it's
    frozen after `.new`, so `@result_ast ||=` fails with `FrozenError` if
    we try to get `#errors`, for example
    
    Also fix the typo `param` -> `option`
```

Unfortunately can't invent not-noisy benchmark, so tested by this:
```ruby
VALID_INPUT = {
  name: "John",
  age: 20,
  email: "john@doe.com"
}.freeze

INVALID_INPUT = {
  name: :John,
  age: "17",
  email: nil
}.freeze

MESSAGE_COMPILER = PersonSchema.(VALID_INPUT.dup).message_compiler.freeze

def result(input)
  Dry::Schema::Result.method(:new).super_method.call(input.dup, message_compiler: MESSAGE_COMPILER)
end

key_map = Dry::Schema::KeyMap.new([[:name, :age, :email]])
coercer =  Dry::Schema::KeyCoercer.new(key_map, &:to_sym)
step = Dry::Schema::Step.new(executor: coercer, type: :core, name: :key_coercer)

Benchmark.ips do |x|
  x.config(stats: :bootstrap, confidence: 95)

  x.report("valid input") { step.(result(VALID_INPUT)) }
  x.report("invalid input") { step.(result(INVALID_INPUT)) }
  x.compare!
end
```
Before:
```
Warming up --------------------------------------
         valid input    11.959k i/100ms
       invalid input    12.077k i/100ms
Calculating -------------------------------------
         valid input    113.668k (± 1.1%) i/s -    574.032k in   5.059266s
       invalid input    122.840k (± 1.2%) i/s -    615.927k in   5.027367s
                   with 95.0% confidence

Comparison:
       invalid input:   122840.2 i/s
         valid input:   113668.2 i/s - 1.08x  (± 0.02) slower
                   with 95.0% confidence
```
After:
```
Warming up --------------------------------------
         valid input    11.004k i/100ms
       invalid input    12.469k i/100ms
Calculating -------------------------------------
         valid input    117.272k (± 1.2%) i/s -    594.216k in   5.077576s
       invalid input    114.757k (± 1.1%) i/s -    573.574k in   5.007477s
                   with 95.0% confidence

Comparison:
         valid input:   117272.0 i/s
       invalid input:   114757.4 i/s - same-ish: difference falls within error
                   with 95.0% confidence
```

In theory both valid & invalid inputs should be the same, since we're not generating error messages, but  there is variation. The same with before/after comparison, results are comparable, but different runs show `115000 i/s ± 10%`.

dry-v specs are passing.