@ojab 2021-06-10
BTW, is there any easy way to run rubocop locally like it run on CI?