@solnic 2021-06-11
Thanks, I appreciate that you explained things in the comments. This is a bold refactoring but again, in the test suite we trust.